refactoring :: BookmarksDomainInterface Transfer 나눔
