+++ Projects/Domain/BookmarksDomain/Interface/DTO/Transfer/BookmarkListTransfer.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+public extension BookmarkListResponseDTO {
+    func toDomain() -> BookmarkListEntity {
+        BookmarkListEntity(bookmarks: bookmarks.map { $0.toDomain() })
+    }
+}
+
+public extension BookmarkResponseDTO {
+    func toDomain() -> BookmarkEntity {
+        BookmarkEntity(
+            companyName: companyName,
+            recruitmentID: recruitmentID,
+            createdAt: createdAt
+        )
+    }
+}
