+++ Projects/Domain/BookmarksDomain/Interface/DataSource/RemoteBookmarksDataSource.swift
@@ -2,6 +2,6 @@ import Combine
 import Foundation
 
 public protocol RemoteBookmarksDataSource {
-    func fetchBookmarkList() -> AnyPublisher<BookmarksListEntity, Error>
+    func fetchBookmarkList() -> AnyPublisher<BookmarkListEntity, Error>
     func bookmark(id: Int) -> AnyPublisher<Void, Error>
 }
