+++ Projects/Domain/BookmarksDomain/Interface/Entity/BookmarkEntity.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+public struct BookmarkEntity: Equatable {
+    public let companyName: String
+    public let recruitmentID: Int
+    public let createdAt: String
+
+    public init(
+        companyName: String,
+        recruitmentID: Int,
+        createdAt: String
+    ) {
+        self.companyName = companyName
+        self.recruitmentID = recruitmentID
+        self.createdAt = createdAt
+    }
+}
