+++ Projects/Domain/BookmarksDomain/Interface/Repository/BookmarksRepository.swift
@@ -1,6 +1,6 @@
 import Combine
 
 public protocol BookmarksRepository {
-    func fetchBookmarkList() -> AnyPublisher<BookmarksListEntity, Error>
+    func fetchBookmarkList() -> AnyPublisher<BookmarkListEntity, Error>
     func bookmark(id: Int) -> AnyPublisher<Void, Error>
 }
