+++ Projects/Domain/BookmarksDomain/Interface/UseCase/FetchBookmarkListUseCase.swift
@@ -1,5 +1,5 @@
 import Combine
 
 public protocol FetchBookmarkListUseCase {
-    func execute() -> AnyPublisher<BookmarksListEntity, Error>
+    func execute() -> AnyPublisher<BookmarkListEntity, Error>
 }
