+++ Projects/Domain/BookmarksDomain/Sources/DataSource/RemoteBookmarksDataSourceImpl.swift
@@ -4,8 +4,8 @@ import Combine
 
 public final class RemoteBookmarksDataSourceImpl: BaseRemoteDataSource<BookmarksDomain>,
                                                   RemoteBookmarksDataSource {
-    public func fetchBookmarkList() -> AnyPublisher<BookmarksListEntity, Error> {
-        request(.fetchBookmarkList, dto: BookmarksListResponseDTO.self)
+    public func fetchBookmarkList() -> AnyPublisher<BookmarkListEntity, Error> {
+        request(.fetchBookmarkList, dto: BookmarkListResponseDTO.self)
             .map { $0.toDomain() }
             .eraseToAnyPublisher()
     }
